Fix the not a git repo error when running the executable
'fatal: Not a git repository (or any of the parent directories): .git'
was caused by the git ls-files in the gemspec

We were requiring it the executable. The files for the gemspec are now
manually loaded. The chdir is there so that we always load the correct
files. (the current dir could be bogus)