Prefix monorepo applications
If there is a repository which contains multiple Ruby projects with
separate `Gemfile`s, currently the project is displayed as the directory
name.

However, most repositories use names that are meaningful within context,
so `lawnmover/starter` is better than `starter`.

This changeset decorates nested applications with the name relative to
the root directory.